Get rid of ioutil.ReadAll()
ioutil.ReadAll() was called with the intention to unmarshal the json of a
Reader. This commit changes the code to use json.Decoder instead, when
the Debug flag is false.
For Debug = true, the behavior remains the same.
There is another instance of ioutil.ReadAll() (in UploadFile)
but it can be worked around and is properly documented.